# fix the allreduce fused bug