<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
Bug fixes

### PR changes
Others

### Describe
Rewrite the logic for allreduce fuse under raw_program_optimizer.

Test performance and precision based on MOE model on **8 A100 GPUs**.
#### Performance gain
| Without fuse throughput | Fuse_num = 8 throughput | Performance gain |
|:-------------------------:|:-------------------------:|:------------------:|
| 5.213 steps / sec | 5.453 steps / sec | +4.6% |

#### Loss curve
![Screen Shot 2021-07-28 at 2 26 41 PM](https://user-images.githubusercontent.com/25279174/127274375-f63c81b8-62cc-4a4f-9307-e2e5c2df6017.png)

#### Loss diff
![Screen Shot 2021-07-28 at 2 29 58 PM](https://user-images.githubusercontent.com/25279174/127274726-17a95841-b573-4201-88ba-9576317255fa.png)
max loss diff: 0.2944
min loss diff: 0.0001

#### How to use
```python
strategy = paddle.distributed.fleet.DistributedStrategy()
strategy.without_graph_optimization = True
strategy.fuse_all_reduce_ops = True
# the following two flags are optional
strategy._calc_comm_same_stream = False    # use calc stream as comm stream or not, default is False
strategy.fuse_grad_size_in_num = 8         # the number of grad in one allreduce_sum op, default is 8
```